requirements.txt
anthropic>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.0
pypdf>=4.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
